requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
Pillow==10.0.0
drf-yasg==1.21.7 
inflection==0.5.1 
packaging==23.1 
pyyaml==6.0.1 
uritemplate==4.1.1
django-phonenumber-field==7.1.0
django-twilio==0.14.3.2 
idna==3.4 
phonenumbers==8.13.19 
requests==2.31.0 
twilio==7.17.0 
urllib3==2.0.4
django-phonenumbers==1.0.1
